requirements.txt
flask
pypdf
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
chromadb==1.0.20
sentence-transformers==5.1.0
google-generativeai==0.8.5
python-dotenv==1.1.1
nest-asyncio==1.6.0